+++ CHANGES.rst
@@ -10,3 +10,4 @@ v2019.1 YYYY-MM-DD
 * Unify the TRS interfaces between the node and the TRSTransform classes, by
   using and exposing a TRSTransform instance inside the node class (GitLab #46)
 * Select polygon mode from GUI, and simplify it (GitLab #47)
+* Toggle visualisation of light cones in wireframe mode from the GUI